Shell-based installer for NGS & NSC
This is based on the installer I wrote in nats-io/nsc in February.

We are not quite pure POSIX, the comment block under the license shows
what's needed/assumed.  If curl or unzip are missing, we will handle it.
If we need to work with something which does not support 'local', then
we can adapt by becoming more unsafe, but it will be a bit ugly.
Uglier.