Installer: Unbreak Alpine in Docker
The `uname -p` command can return `unknown`, even while `arch` returns a real
value.  Even though uname is more widely guaranteed to be present.  So, handle
"unknown" as a special case and try arch(1).  If calling arch fails then we'll
abort and we're no worse off than we were with an "unknown" value.

Make sure that errors inside command substitution sub-shells propagate properly
out, by splitting a printf invocation to use an intermediate variable setting
first; this way, printf(1) won't swallow the $? and mask it, and we can force
an exit if the sub-shell failed.

This repairs an installability regression introduced with the switch to
multi-arch support (for handling arm64); prior to that, we assumed amd64
always.